+++ src/resources/upload-requests.ts
@@ -79,7 +79,13 @@ export class UploadRequests {
   async fulfill(request: UploadRequestFulfillRequest): Promise<UploadFulfillment> {
     const form = new FormData();
     await appendFile(form, request.file, request.filename);
-    return this.httpClient.postForm<UploadFulfillment>(request.uploadUrl, form, {}, false);
+    return this.httpClient.postForm<UploadFulfillment>(
+      request.uploadUrl,
+      form,
+      {},
+      false,
+      'upload_requests.upload'
+    );
   }
 }